requirements.txt
huggingface_hub==0.33.0
langchain==0.3.26
langchain-community==0.3.26
langchain-huggingface==0.3.0
gupload==1.1.0
faiss-gpu-cu12==1.11.0
bertopic==0.17.0
pynndescent==0.5.10
torch @ https://download.pytorch.org/whl/cu124/torch-2.6.0%2Bcu124-cp311-cp311-linux_x86_64.whl
pytrec_eval
json
tqdm